ポイントカードをコンポーネント化
